Vector math: Consolidate existing helpers into a single file.
