+++ shared-data/js/helpers/index.ts
@@ -31,8 +31,7 @@ export * from './wellSets'
 export * from './getAreFlexSlotsAdjacent'
 export * from './getMmFromBottom'
 export * from './getModuleVizDims'
-export * from './getVectorDifference'
-export * from './getVectorSum'
+export * from './vectorMath'
 export * from './getLoadedLabwareDefinitionsByUri'
 export * from './getFixedTrashLabwareDefinition'
 export * from './getOccludedSlotCountForModule'
