+++ shared-data/js/helpers/vectorMath.ts
@@ -1,5 +1,13 @@
 import type { Vector3D } from '../types'
 
+export function getVectorSum(pointA: Vector3D, pointB: Vector3D): Vector3D {
+  return {
+    x: pointA.x + pointB.x,
+    y: pointA.y + pointB.y,
+    z: pointA.z + pointB.z,
+  }
+}
+
 export function getVectorDifference(
   pointA: Vector3D,
   pointB: Vector3D
